requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.2
pip==20.1.1
psycopg2==2.9.1
pytz==2021.1
setuptools==47.1.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
whitenoise==5.3.0
drf-yasg==1.20.0